Changed the order that objects are built in
Old order was:
1. Ensembles/Nodes
2. Probes
3. Connections
4. Learning Rules

New order is:
1. Ensembles/Nodes
2. Non-probe Connections
3. Learning Rules
4. Probes
5. Probe Connections

Probe Connections are the connections that probes use to get their
input.

This is part of an effort to make connections probable (so, for
example, we can see how a connection weight matrix changes over time).
Basically, the objects that a connection connects should be built
before that connection, so that build_connection can extract the
proper signals. In particular, probe connections have to be built
after normal connections if we allow connections to be probed. Probe
connections also have to be built after probes, which was probably
part of the reason for the original order.