Added hierarchical seeding.
Before building, assigns seeds to objects by doing a depth-first search
of the hierarchical network structures. Objects at the same level are
assigned seeds according to the order that they are created in (i.e. the
order their __init__ function are called). For each subnetwork, a new
seed is generated, and the seeds for components of the subnetwork are
based off that.

The effect is that changes to the network generation code will not
necessarily change the seed used to build an object. For example, one
can completely change the contents of a subnetwork without changing the
seeds assigned to objects in the top level network, provided that the
top level objects are still created in the same order.