Refactored SignalDict, added reset tests
The _get_base_array function wasn't really necessary; we just have
to make sure that we're making a copy of the array in init so that
we don't end up modifying it. In reset, we have a guarantee that the
ndarray pointed to by `signal` won't change, so we can just copy in
the original value.

Added tests to test_builder and test_simulator to ensure that
resetting actually works.