SQUASHME: refactor conftest.py and utils/testing.py
The activate_recorder function was one level of indirection too far,
I think. With this, we instead use `dirname is None` to mean
"don't bother recording", rather than explicitly tracking it.

This also makes the Recorder subclasses a bit more flexible by using
properties for things that were previously attributes.

Also added a docstring to the Analytics fixture, as it gets printed
with `py.test --fixtures`.