Make old network inputs available as deprecated.
This commit converts the Product and CircularConvolution network
functions into classes to be able to add properties for the old inputs
that give DeprecationWarnings. Because with the class we are not able
to handle the net argument properly (except with some magic metaclass
stuff that probably is not a good idea), it is made obsolete. Network
arguments like label and seed can now passed as keyword arguments.

I decided for the kwargs approach instead of listing the explicit
arguments because:

* less boilerplate
* the relevant arguments seed and label are easy to remember (so users
  won't have to go back to the Network documentation a lot)
* the argument add_to_container was for me always
  a hassle to remember when subclassing and isn't that relevant to the
  average user
* kwargs enforces to use keyword arguments when instantiating the
  network.
* It is a pretty common idiom (e.g. used in Matplotlib)

I like that this simplified the instantiation in a few tests that
actually were using the net argument.

Addresses #887.