Add test that optimizer does not change result.
Modified the test to reduce repetition with other tests, and to
make it easier to add new tests if we notice things that are
difficult to optimize correctly. Also probing more things, to
ensure that we are actual doing exactly the same thing, and
not just something that is functionally the same.

This uncovered a bug with `sig_replacements`; if a signal is
replaced with a signal that then also gets replaced, we weren't
dealing with that chain of replacements. Fixed the bug in this
commit as well.